==> v2-refined/Part1-data_retrival.ipynb <==
##Start

#### Code Summary
The code connects to the National Library of Medicine's FTP server to download 'mutation2pubtatorcentral.gz.' It retrieves the last modified date of the file and checks if it already exists locally. If the local file is older, it downloads the updated file. After download, it displays the last modified date and the file size. Finally, it exits the FTP session.

This script helps ensure that the local copy of the file is always up to date by comparing timestamps. It successfully downloaded 'mutation2pubtatorcentral.gz' from the FTP server, confirming that the process was completed.

import ftplib
import os
from datetime import datetime

# NLM-FTP server details
ftp_url = 'ftp.ncbi.nlm.nih.gov'
ftp_directory = '/pub/lu/PubTatorCentral/'
filename = 'mutation2pubtatorcentral.gz'

try:
    # Start the FTP session
    ftp = ftplib.FTP(ftp_url)
    ftp.login()
    ftp.cwd(ftp_directory)

    # Get the last modified date of the file
    last_modified = ftp.sendcmd(f"MDTM {filename}")
    last_modified_datetime = datetime.strptime(last_modified[4:], "%Y%m%d%H%M%S")
    print(f"Last Modified Date of {filename}: {last_modified_datetime}")

    # Check if the file already exists locally
    if os.path.exists(filename):
        local_last_modified = os.path.getmtime(filename)
        if local_last_modified >= last_modified_datetime.timestamp():
            print(f"{filename} is already up to date.")
            exit()

    with open(filename, 'wb') as local_file:
        ftp.retrbinary(f"RETR {filename}", local_file.write)

    file_size = os.path.getsize(filename)
    print(f"Downloaded {filename} with a size of {file_size / (1024 * 1024):.2f} MB")

except Exception as e:
    print(f"Error encountered: {e}")

finally:
    ftp.quit()

#END
